% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_a1}
\alias{cal_a1}
\title{calculate a1 from correlations}
\usage{
cal_a1(rxmo, rmomu, rxmu)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rmomu}{correlation between Mo and Mu}

\item{rxmu}{correlation between X and Mu}
}
\value{
the value of a1
}
\description{
calculate a1 as a function of correlations
}
