% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_a2}
\alias{cal_a2}
\title{calculate a2 from correlations}
\usage{
cal_a2(rxmu)
}
\arguments{
\item{rxmu}{correlation between X and Mu}
}
\value{
the value of a2
}
\description{
calculate a2 as a function of correlations
}
