% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_b2}
\alias{cal_b2}
\title{calculate b2 from correlations}
\usage{
cal_b2(rxmo, rxmu, rmomu, rxy, rymo, rymu)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rymu}{correlation between Y and Mu}
}
\value{
the value of b2
}
\description{
calculate b2 as a function of correlations
}
