% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_bias_a1}
\alias{cal_bias_a1}
\title{calculate bias of a1 from correlations}
\usage{
cal_bias_a1(rxmo, rxmu, rmomu)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}
}
\value{
the value of bias of a1
}
\description{
calculate bias of a1 as a function of correlations
}
