% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_bias_a1_perc}
\alias{cal_bias_a1_perc}
\title{calculate percentage of bias for a1 from correlations}
\usage{
cal_bias_a1_perc(rxmo, rxmu, rmomu)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}
}
\value{
the percentage of bias for a1
}
\description{
calculate percentage of bias for a1 a function of correlations
}
