% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_bias_b1_perc}
\alias{cal_bias_b1_perc}
\title{calculate percentage of bias for b1 from correlations}
\usage{
cal_bias_b1_perc(rxmo, rxmu, rmomu, rxy, rymo, rymu)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rymu}{correlation between Y and Mu}
}
\value{
the percentage of bias for b1
}
\description{
calculate percentage of bias for b1 a function of correlations
}
