% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_bias_c}
\alias{cal_bias_c}
\title{calculate bias of c from correlations}
\usage{
cal_bias_c(rxmo, rxmu, rmomu, rxy, rymo, rymu)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rymu}{correlation between Y and Mu}
}
\value{
the value of bias of c
}
\description{
calculate bias of c as a function of correlations
}
