% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cal_k}
\alias{cal_k}
\title{calculate k from correlations}
\usage{
cal_k(rxmo, rxmu, rmomu)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}
}
\value{
the value of k
}
\description{
calculate k as a function of correlations
}
