% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_constraint.R
\name{check_input}
\alias{check_input}
\title{check if input is valid}
\usage{
check_input(a)
}
\arguments{
\item{a}{each element of the input/correlation}
}
\value{
True if the input is valid otherwise False
}
\description{
check whether the user gives valid correlations
}
