% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmed_ind_ME.R
\name{conmed_ind_ME}
\alias{conmed_ind_ME}
\title{Perform sensitivity analysis for mediation analysis}
\usage{
conmed_ind_ME(
  est_eff_a,
  std_err_a,
  est_eff_b,
  std_err_b,
  n_obs,
  n_covariates_a = 0,
  n_covariates_b = 1,
  rel_M = 1,
  rel_Y = 1,
  rel_X = 1,
  sd_X,
  sd_M,
  sd_Y,
  R2_a,
  R2_b,
  alpha = 0.05,
  tails = 2,
  nu = 0
)
}
\arguments{
\item{est_eff_a}{the estimated effect for path a (from treatment to mediator)}

\item{std_err_a}{the standard error of the estimate of path a (from treatment to mediator)}

\item{est_eff_b}{the estimated effect for path b (from mediator to outcome)}

\item{std_err_b}{the standard error of the estimate of path b (from mediator to outcome)}

\item{n_obs}{the number of observations in the sample}

\item{n_covariates_a}{the number of covariates in the regression model for estimating path a (from treatment to mediator), including the}

\item{n_covariates_b}{the number of covariates in the regression model for estimating path b (from mediator to outcome)}

\item{rel_M}{the reliability level of the mediator}

\item{rel_Y}{the reliability level of the outcome}

\item{rel_X}{the reliability level of the treatment}

\item{sd_X}{the standard deviation of the treatment}

\item{sd_M}{the standard deviation of the mediator}

\item{sd_Y}{the standard deviation of the outcome}

\item{R2_a}{the R square of the regression model that estiamtes the a pathway (from treatment to mediator)}

\item{R2_b}{the R square of the regression model that estiamtes the b pathway (from mediator to outcome)}

\item{alpha}{probability of rejecting the null hypothesis (defaults to 0.05)}

\item{tails}{integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)}
}
\value{
prints the ITCV to invalidate/sustain the inference & show the impact curve
}
\description{
This command calculates the impact of an omitted variable necessary to invalidate/sustain an inference for a mediation effect, allowing for potential measurment error.
}
\examples{
conmed_ind_ME(est_eff_a = 0.171, std_err_a = 0.087, est_eff_b = 0.387, std_err_b = 0.077, n_obs = 123, rel_M = 0.9, rel_Y = 0.9, sd_X = 1, sd_M = 1.054093, sd_Y = 1.054093, R2_a = 0.02941117, R2_b = 0.1685979)
}
