% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmed_plot.R
\name{conmed_plot_a1b1}
\alias{conmed_plot_a1b1}
\title{sensitivity plot for estimates with regards to all correlations regarding Mu}
\usage{
conmed_plot_a1b1(rxmo = -2, rxy = -2, rymo = -2, nobs = nobs, conflevel = 0.95)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{nobs}{number of observations in the sample or sample size}

\item{conflevel}{confidence interval level, default is 0.95}
}
\value{
plot for how rxmu rymu rmomu affects the estimates and confidence interval
}
\description{
give the sensitivity plot for how the correlation between Y and Mu affects the estimates and confidence interval
}
