% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_constraint.R
\name{constrain_standardized}
\alias{constrain_standardized}
\title{check the standardization constraint}
\usage{
constrain_standardized(a1, a2, b1, b2, k, c)
}
\arguments{
\item{a1}{parameter for path X to Mo}

\item{a2}{parameter for path X to Mu}

\item{b1}{parameter for path Mo to Y}

\item{b2}{parameter for path Mu to Y}

\item{k}{parameter for path Mu to Mo}

\item{c}{parameter for path X to Y}
}
\value{
True if the constraint is fulfilled otherwise False
}
\description{
check if the combination of correlations fulfil the standardization constraint
}
