% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_output_print.R
\name{output_print}
\alias{output_print}
\title{Function to output printed text}
\usage{
output_print(
  focus,
  ind,
  est_eff,
  beta_threshhold,
  nu,
  obs_r,
  critical_r,
  r_con,
  itcv,
  alpha
)
}
\arguments{
\item{focus}{what is the scenario}

\item{ind}{estimated indirect effect}

\item{est_eff}{estimated effect}

\item{nu}{null hypothesis}

\item{obs_r}{observed r value}

\item{critical_r}{critical r value}

\item{r_con}{correlaton for achieving ITCV}

\item{itcv}{ITCV}

\item{alpha}{alpha value}

\item{beta_threshold}{threshold for coefficient}
}
\value{
print output
}
\description{
Function to output printed text
}
