% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_output_print.R
\name{output_print_both}
\alias{output_print_both}
\title{Function to output printed text for sustaining both a and b}
\usage{
output_print_both(
  ind,
  est_eff_a,
  est_eff_b,
  beta_threshold_a,
  beta_threshold_b,
  nu,
  obs_r_a,
  obs_r_b,
  critical_r_a,
  critical_r_b,
  r_con_a,
  r_con_b,
  itcv_a,
  itcv_b,
  alpha
)
}
\arguments{
\item{ind}{estimated indirect effect}

\item{est_eff_a}{estimated effect for a}

\item{est_eff_b}{estimated effect for b}

\item{beta_threshold_a}{threshold for coefficient (a)}

\item{beta_threshold_b}{threshold for coefficient (b)}

\item{obs_r_a}{observed r value for a pathway}

\item{obs_r_b}{observed r value for b pathway}

\item{critical_r_a}{critical r value for a pathway}

\item{critical_r_b}{critical r value for b pathway}

\item{r_con_a}{correlaton for achieving ITCV for a pathway}

\item{r_con_b}{correlaton for achieving ITCV for b pathway}

\item{itcv_a}{ITCV for a pathway}

\item{itcv_b}{ITCV for b pathway}

\item{alpha}{alpha value}
}
\value{
print output
}
\description{
Function to output printed text for sustaining both a and b
}
