% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_output_plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{plot impact curve}
\usage{
plot_curve(focus, obs_r, critical_r, r_con)
}
\arguments{
\item{focus}{what is the scenario}

\item{obs_r}{observed r value}

\item{critical_r}{critical r value}

\item{r_con}{correlaton for achieving ITCV}
}
\value{
plot for impact curve
}
\description{
plot impact curve
}
