% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataresult.R
\name{rmomu_data}
\alias{rmomu_data}
\title{how the percentage of bias with regards to correlation between Mo and Mu}
\usage{
rmomu_data(rxmo = -2, rxy = -2, rymo = -2, rxmu = -2, rymu = -2)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rymu}{correlation between Y and Mu}
}
\value{
result for how rmomu affects the percentage of bias
}
\description{
give the result for how the correlation between Mo and Mu affects the percentage of bias
}
