% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresult.R
\name{rmomu_plot}
\alias{rmomu_plot}
\title{sensitivity plot for estimates and confidence interval with regards to correlation between Mo and Mu}
\usage{
rmomu_plot(rxmo = -2, rxy = -2, rymo = -2, rxmu = -2, rymu = -2,
  nobs = 0, labelest = "indirect1", conflevel = 0.95)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rymu}{correlation between Y and Mu}

\item{nobs}{number of observations in the sample or sample size}

\item{labelest}{label for which estimate the figure is for: indirect1, a1, b1 or c, default is indirect1}

\item{conflevel}{confidence interval level, default is 0.95}
}
\value{
plot for how rmomu affects the estimates and confidence interval
}
\description{
give the sensitivity plot for how the correlation between Mo and Mu affects the estimates and confidence interval
}
