% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataresult.R
\name{rxmu_data}
\alias{rxmu_data}
\title{how the estimate and confidence interval change with regards to correlation between X and Mu}
\usage{
rxmu_data(
  rxmo = -2,
  rxy = -2,
  rymo = -2,
  rmomu = -2,
  rymu = -2,
  nobs = 0,
  conflevel = 0.95
)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rmomu}{correlation between Mo and Mu}

\item{rymu}{correlation between Y and Mu}

\item{nobs}{number of observations in the sample or sample size}

\item{conflevel}{confidence interval level, default is 0.95}
}
\value{
result for how rxmu affects the estimates
}
\description{
give the result for how the estimate and confidence interval change with regards to correlation between X and Mu
}
