% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_a1b1.R
\name{rxmu_plot_a1b1}
\alias{rxmu_plot_a1b1}
\title{sensitivity plot for estimates with regards to correlation between X and Mu}
\usage{
rxmu_plot_a1b1(
  rxmo = -2,
  rxy = -2,
  rymo = -2,
  nobs = nobs,
  conflevel = 0.95,
  specifyunob = 0,
  rmomu = -2,
  rymu = -2
)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{nobs}{number of observations in the sample or sample size}

\item{conflevel}{confidence interval level, default is 0.95}

\item{specifyunob}{whether the user wants to specify two unobserved correlations default is 0, meaning no}

\item{rmomu}{correlation between Mo and Mu}

\item{rymu}{correlation between Y and Mu}
}
\value{
plot for how rxmu affects the estimates and confidence interval
}
\description{
give the sensitivity plot for how the correlation between X and Mu affects the estimates and confidence interval
regarding a1, b1, and a1*b1
}
