% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresult.R
\name{rxmu_plot_auxi}
\alias{rxmu_plot_auxi}
\title{auxillary function for sensitivity plot for estimates with regards to correlation between X and Mu}
\usage{
rxmu_plot_auxi(rxmo = -2, rxy = -2, rymo = -2, rmomu = -2,
  rymu = -2, nobs = 0, labelest = "indirect1", conflevel = 0.95)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rmomu}{correlation between Mo and Mu}

\item{rymu}{correlation between Y and Mu}

\item{nobs}{number of observations in the sample or sample size}

\item{labelest}{label for which estimate the figure is for: indirect1, a1, b1 or c, default is indirect1}

\item{conflevel}{confidence interval level, default is 0.95}
}
\value{
plot for how rxmu affects the estimates and confidence interval
}
\description{
give the sensitivity plot for how the correlation between X and Mu affects the estimates and confidence interval
}
