% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresult_perc.R
\name{rxmu_plot_perc}
\alias{rxmu_plot_perc}
\title{sensitivity plot for percentage of bias with regards to correlation between X and Mu}
\usage{
rxmu_plot_perc(rxmo = -2, rxy = -2, rymo = -2, rmomu = -2,
  rymu = -2)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rmomu}{correlation between Mo and Mu}

\item{rymu}{correlation between Y and Mu}
}
\value{
plot for how rxmu affects the percentage of bias
}
\description{
give the sensitivity plot for how the correlation between X and Mu affects the percentage of bias
}
