% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataresult_perc.R
\name{rymu_data_perc}
\alias{rymu_data_perc}
\title{how the percentage of bias with regards to correlation between Y and Mu}
\usage{
rymu_data_perc(rxmo = -2, rxy = -2, rymo = -2, rxmu = -2,
  rmomu = -2)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}
}
\value{
result for how rymu affects the percentage of bias
}
\description{
give the result for how the correlation between Y and Mu affects the percentage of bias
}
