% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresult.R
\name{rymu}
\alias{rymu}
\title{sensitivity plot for percentage of bias with regards to correlation between Y and Mu}
\usage{
rymu(rxmo = -2, rxy = -2, rymo = -2, rxmu = -2, rmomu = -2)
}
\arguments{
\item{rxmo}{correlation between X and Mo}

\item{rxy}{correlation between X and Y}

\item{rymo}{correlation between Y and Mo}

\item{rxmu}{correlation between X and Mu}

\item{rmomu}{correlation between Mo and Mu}
}
\value{
plot for how rymu affects the percentage of bias
}
\description{
give the sensitivity plot for how the correlation between Y and Mu affects the percentage of bias
}
